feat(types): add IConfig and event interfaces for improved type safety; restructure type exports
