+++ src/types/index.ts
@@ -1,12 +1,17 @@
 import discord from "discord.js";
 
 import CommandLogger from "../core/command/logger";
-import { ILogger } from "./logger";
 
-export * from "./logger";
-export * from "./ticket";
+import { ILogger } from "./logger";
+import { IConfig } from "./config";
+import { SlashCommand, Command } from "./events";
+
 export * from "./db";
 export * from "./ai";
+export * from "./logger";
+export * from "./ticket";
+export * from "./events";
+export * from "./config";
 
 
 declare global {
@@ -29,118 +34,4 @@ declare module "discord.js" {
         cmdLogger: CommandLogger;
         config: IConfig;
     }
-}
-
-export interface SlashCommand {
-    data: discord.SlashCommandBuilder | discord.SlashCommandSubcommandsOnlyBuilder | discord.SlashCommandOptionsOnlyBuilder;
-    modal?: (
-        interaction: discord.ModalSubmitInteraction<discord.CacheType>
-    ) => void;
-    userPerms?: Array<discord.PermissionResolvable>;
-    botPerms?: Array<discord.PermissionResolvable>;
-    cooldown?: number;
-    owner?: boolean;
-    premium?: boolean;
-    execute: (
-        interaction: discord.ChatInputCommandInteraction,
-        client: discord.Client
-    ) => void;
-    autocomplete?: (
-        interaction: discord.AutocompleteInteraction,
-        client: discord.Client
-    ) => void;
-}
-
-export interface Command {
-    name: string;
-    description: string;
-    userPerms?: Array<discord.PermissionResolvable>;
-    botPerms?: Array<discord.PermissionResolvable>;
-    cooldown?: number;
-    owner?: boolean;
-    premium?: boolean;
-    execute: (
-        client: discord.Client,
-        message: discord.Message,
-        args: Array<string>
-    ) => void;
-}
-
-export interface BotPresence {
-    name: string;
-    type: discord.ActivityType;
-}
-
-export interface IConfig {
-    bot: {
-        owners: Array<string>;
-        support: {
-            link: string;
-        };
-        presence: {
-            enabled: boolean;
-            status: string;
-            interval: number;
-            activity: Array<BotPresence>;
-        };
-        command: {
-            prefix: string;
-            disable_message: boolean;
-            cooldown_message: string;
-            register_specific_commands: {
-                enabled: boolean;
-                commands: Array<string>;
-            };
-        };
-        log: {
-            command: string;
-            server: string;
-        };
-    };
-    embed: {
-        color: {
-            default: discord.ColorResolvable | null;
-            error: discord.ColorResolvable | null;
-            success: discord.ColorResolvable | null;
-            warning: discord.ColorResolvable | null;
-        };
-    };
-    ai: {
-        chatbot: {
-            embedding: {
-                model: string;
-            };
-        };
-    };
-    ticket: {
-        default: {
-            button: {
-                label: string;
-                emoji: string;
-                style: string;
-                embed_title: string;
-                embed_description: string;
-            };
-            category: {
-                name: string;
-                description: string;
-                emoji: string;
-            };
-            message: {
-                welcome_message: string;
-                close_message: string;
-            };
-            select_menu: {
-                placeholder: string;
-                embed_title: string;
-                embed_description: string;
-            };
-        };
-    };
-};
-
-export interface BotEvent {
-    name: string;
-    once?: boolean | false;
-    execute: (...args: any[]) => void | Promise<void>;
 }
